Rename `mlgame.exception` to `mlgame.exceptions`
